rename advancedUserTargeting -> advancedTargeting
