+++ apps/web/app/(app)/environments/[environmentId]/(peopleAndSegments)/segments/components/EditSegmentModal.tsx
@@ -2,7 +2,7 @@
 
 import { UserGroupIcon } from "@heroicons/react/24/solid";
 
-import SegmentSettingsTab from "@formbricks/ee/advancedUserTargeting/components/SegmentSettings";
+import SegmentSettingsTab from "@formbricks/ee/advancedTargeting/components/SegmentSettings";
 import { TActionClass } from "@formbricks/types/actionClasses";
 import { TAttributeClass } from "@formbricks/types/attributeClasses";
 import { TSegment, TSegmentWithSurveyNames } from "@formbricks/types/segment";
